Reject messages from clients that are connected to the wrong server
This will make them eventually timeout and try a new server instead of
trying to use a server that they aren't supposed to be using.

This can happen on races of clients connecting to multiple servers
due to a sequence of timeouts (frequent on high-load scenarios) and
the ConnectRequest/AcceptRequests ending up on the Leader at the
wrong order.

Example:

1) Client A tries to connect to Server X and times out
2) Client A tries to connect to Server Y and succeeds
3) Leader receives AcceptRequest from Server Y and processes/publishes it
4) Leader receives AcceptRequest from Server X and processes/publishes it

After this, Client A is connected to Server Y but the cluster believes
it should be connected to server X. This causes some inconsistency
checks that eventually leads to the Client A becoming useless and having
all its operations timeout. Its internal session state won't receive
publishes because Server Y won't send them (its connection was closed
once it received the message that Client A is now connected to Server
X).